requirements.txt
fastapi
uvicorn
sentry-sdk[fastapi]
tortoise-orm[asyncpg]
python-dotenv